refactor(advisor): share advisor utilities
