+++ tools/advisors/json.mts
@@ -1,0 +1,77 @@
+// SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export function extractJson(text: string, rawPath: string, tag: string, label = "advisor output"): unknown {
+  const trimmed = text.trim();
+  const candidates = [trimmed, fenced(trimmed), tagged(trimmed, tag), balancedObject(trimmed)].filter(
+    (candidate): candidate is string => Boolean(candidate),
+  );
+
+  for (const candidate of candidates) {
+    try {
+      return JSON.parse(candidate);
+    } catch {
+      // Try next candidate.
+    }
+  }
+  throw new Error(`Could not parse JSON from ${label}; see ${rawPath}`);
+}
+
+export function enumValue<T extends readonly string[]>(value: unknown, allowed: T, fallback: T[number]): T[number] {
+  return typeof value === "string" && allowed.includes(value) ? value : fallback;
+}
+
+export function recordItems(value: unknown): Record<string, unknown>[] {
+  return Array.isArray(value) ? value.filter(isRecord) : [];
+}
+
+export function isRecord(value: unknown): value is Record<string, unknown> {
+  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
+}
+
+export function stringArray(value: unknown): string[] {
+  return Array.isArray(value) ? value.filter((item): item is string => typeof item === "string" && item.trim().length > 0).map((item) => item.trim()) : [];
+}
+
+export function stringOrUndefined(value: unknown): string | undefined {
+  return typeof value === "string" && value.trim() ? value.trim() : undefined;
+}
+
+export function stringOrDefault(value: unknown, fallback: string): string {
+  return stringOrUndefined(value) || fallback;
+}
+
+export function dropUndefinedValues<T extends Record<string, unknown>>(object: T): T {
+  return Object.fromEntries(Object.entries(object).filter(([, value]) => value !== undefined)) as T;
+}
+
+export function getPath<T>(value: unknown, pathParts: (string | number)[]): T | undefined {
+  let current: unknown = value;
+  for (const part of pathParts) {
+    if (typeof part === "number") {
+      if (!Array.isArray(current)) return undefined;
+      current = current[part];
+      continue;
+    }
+    if (!isRecord(current)) return undefined;
+    current = current[part];
+  }
+  return current as T | undefined;
+}
+
+function fenced(text: string): string | undefined {
+  const match = text.match(/```(?:json)?\s*([\s\S]*?)```/i);
+  return match?.[1]?.trim();
+}
+
+function tagged(text: string, tag: string): string | undefined {
+  const match = text.match(new RegExp(`<${tag}>\\s*([\\s\\S]*?)\\s*</${tag}>`, "i"));
+  return match?.[1]?.trim();
+}
+
+function balancedObject(text: string): string | undefined {
+  const start = text.indexOf("{");
+  const end = text.lastIndexOf("}");
+  if (start === -1 || end === -1 || end <= start) return undefined;
+  return text.slice(start, end + 1);
+}
