+++ tools/pr-review-advisor/analyze.mts
@@ -2,26 +2,20 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { execFileSync } from "node:child_process";
 import fs from "node:fs";
 import path from "node:path";
 import { pathToFileURL } from "node:url";
 
-import type { AgentSessionEvent } from "@earendil-works/pi-coding-agent";
-import {
-  AuthStorage,
-  createAgentSession,
-  DefaultResourceLoader,
-  ModelRegistry,
-  SessionManager,
-  SettingsManager,
-} from "@earendil-works/pi-coding-agent";
+import { getChangedFiles, getCommits, getDiff, getDiffStat, getHeadSha, gitOutput } from "../advisors/git.mts";
+import { githubGraphql, githubRest, githubRestPaginated } from "../advisors/github.mts";
+import { advisorArtifactPaths, parseArgs, parsePositiveInt, readJson, writeJson, type AdvisorArtifactPaths } from "../advisors/io.mts";
+import { enumValue, extractJson, getPath, isRecord, recordItems, stringArray, stringOrDefault, stringOrUndefined } from "../advisors/json.mts";
+import { DEFAULT_ADVISOR_MODEL, DEFAULT_ADVISOR_PROVIDER, type RunAdvisorResult, runReadOnlyAdvisor } from "../advisors/session.mts";
 
 const root = process.cwd();
-const ADVISOR_PROVIDER = "openai";
-const ADVISOR_MODEL = "openai/openai/gpt-5.5";
-const READ_ONLY_TOOLS = ["read", "grep", "find", "ls"];
-const ZERO_COST = { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 };
+const ADVISOR_PROVIDER = DEFAULT_ADVISOR_PROVIDER;
+const ADVISOR_MODEL = DEFAULT_ADVISOR_MODEL;
+const ADVISOR_CREDENTIAL_ENV = ["PR", "REVIEW", "ADVISOR", "API", "KEY"].join("_");
 const SECURITY_CATEGORIES = [
   "Secrets and Credentials",
   "Input Validation and Data Sanitization",
@@ -60,8 +54,6 @@ const E2E_STATUS_VERDICTS = ["ok", "missing", "ambiguous", "not_found"] as const
 const ACCEPTANCE_STATUSES = ["met", "partial", "missing", "unknown"] as const;
 const SECURITY_VERDICTS = ["pass", "warning", "fail"] as const;
 
-type ParsedArgs = Record<string, string | undefined>;
-type AdvisorProviderConfig = Parameters<ModelRegistry["registerProvider"]>[1];
 type Confidence = (typeof CONFIDENCES)[number];
 type SummaryRecommendation = (typeof SUMMARY_RECOMMENDATIONS)[number];
 type GateStatusName = (typeof GATE_STATUSES)[number];
@@ -71,19 +63,7 @@ type E2eStatusVerdict = (typeof E2E_STATUS_VERDICTS)[number];
 type AcceptanceStatus = (typeof ACCEPTANCE_STATUSES)[number];
 type SecurityVerdict = (typeof SECURITY_VERDICTS)[number];
 
-type RunAdvisorResult = {
-  text: string;
-  raw: string;
-};
-
-type ArtifactPaths = {
-  prompt: string;
-  raw: string;
-  result: string;
-  finalResult: string;
-  summary: string;
-  sessionHtml: string;
-};
+type ArtifactPaths = AdvisorArtifactPaths;
 
 type ReviewMetadata = {
   baseRef: string;
@@ -197,29 +177,11 @@ type LinkedIssue = {
   fetchError?: string;
 };
 
-const ADVISOR_PROVIDER_CONFIG: AdvisorProviderConfig = {
-  api: "openai-completions",
-  baseUrl: "https://inference-api.nvidia.com/v1",
-  models: [advisorModel(ADVISOR_MODEL, "GPT-5.5", 256000, 32768, true, ["text", "image"])],
-  ["api" + "Key"]: "PR_REVIEW_ADVISOR_API_KEY",
-} as AdvisorProviderConfig;
-
 if (process.argv[1] && import.meta.url === pathToFileURL(process.argv[1]).href) {
   main().catch((error: unknown) => {
     console.error(error instanceof Error ? error.message : String(error));
     process.exit(1);
   });
-}
-
-function advisorModel(
-  id: string,
-  name: string,
-  contextWindow: number,
-  maxTokens: number,
-  reasoning: boolean,
-  input: ("text" | "image")[],
-): NonNullable<AdvisorProviderConfig["models"]>[number] {
-  return { id, name, reasoning, input, cost: ZERO_COST, contextWindow, maxTokens };
 }
 
 async function main(): Promise<void> {
@@ -259,7 +221,7 @@ async function main(): Promise<void> {
   logProgress(`Launching PR review advisor SDK: provider=${ADVISOR_PROVIDER} model=${ADVISOR_MODEL}`);
   let sdkResult: RunAdvisorResult | undefined;
   try {
-    sdkResult = await runAdvisor({
+    sdkResult = await runReadOnlyAdvisor({
       cwd: root,
       prompt,
       systemPrompt,
@@ -268,6 +230,9 @@ async function main(): Promise<void> {
       timeoutMs,
       heartbeatMs,
       maxCaptureBytes,
+      credentialEnv: ADVISOR_CREDENTIAL_ENV,
+      logPrefix: "pr-review-advisor",
+      logProgress,
     });
     fs.writeFileSync(artifacts.raw, sdkResult.raw);
   } catch (error: unknown) {
@@ -279,7 +244,7 @@ async function main(): Promise<void> {
 
   let result: ReviewAdvisorResult;
   try {
-    result = normalizeReviewResult(extractJson(sdkResult.text || sdkResult.raw, artifacts.raw), metadata);
+    result = normalizeReviewResult(extractJson(sdkResult.text || sdkResult.raw, artifacts.raw, "pr_review_advisor_json", "PR review advisor output"), metadata);
   } catch (error: unknown) {
     writeFailure(error instanceof Error ? error.message : String(error));
     process.exit(1);
@@ -293,14 +258,7 @@ async function main(): Promise<void> {
 }
 
 function artifactPaths(outDir: string): ArtifactPaths {
-  return {
-    prompt: path.join(outDir, "pr-review-advisor-prompt.md"),
-    raw: path.join(outDir, "pr-review-advisor-raw-output.txt"),
-    result: path.join(outDir, "pr-review-advisor-result.json"),
-    finalResult: path.join(outDir, "pr-review-advisor-final-result.json"),
-    summary: path.join(outDir, "pr-review-advisor-summary.md"),
-    sessionHtml: path.join(outDir, "pr-review-advisor-session.html"),
-  };
+  return advisorArtifactPaths(outDir, "pr-review-advisor");
 }
 
 function writeUnavailableArtifacts(
@@ -321,269 +279,8 @@ function writeUnavailableArtifacts(
   }
 }
 
-function writeJson(filePath: string, value: unknown): void {
-  fs.writeFileSync(filePath, `${JSON.stringify(value, null, 2)}\n`);
-}
-
 function logProgress(message: string): void {
   console.log(`[pr-review-advisor] ${new Date().toISOString()} ${message}`);
-}
-
-function parsePositiveInt(value: string | undefined, fallback: number): number {
-  const parsed = Number.parseInt(value ?? "", 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : fallback;
-}
-
-async function runAdvisor(options: {
-  cwd: string;
-  prompt: string;
-  systemPrompt: string;
-  configDir: string;
-  htmlExportPath: string;
-  timeoutMs: number;
-  heartbeatMs: number;
-  maxCaptureBytes: number;
-}): Promise<RunAdvisorResult> {
-  fs.mkdirSync(options.configDir, { recursive: true });
-  const { authStorage, modelRegistry } = prepareAdvisorConfig();
-  const model = modelRegistry.find(ADVISOR_PROVIDER, ADVISOR_MODEL);
-  if (!model || !modelRegistry.hasConfiguredAuth(model)) {
-    throw new Error(`Could not configure advisor model ${ADVISOR_MODEL}`);
-  }
-
-  const settingsManager = SettingsManager.inMemory({
-    compaction: { enabled: false },
-    retry: { enabled: true, maxRetries: 2 },
-  });
-  const resourceLoader = new DefaultResourceLoader({
-    cwd: options.cwd,
-    agentDir: options.configDir,
-    settingsManager,
-    noExtensions: true,
-    noSkills: true,
-    noPromptTemplates: true,
-    noThemes: true,
-    noContextFiles: true,
-    systemPromptOverride: () => options.systemPrompt,
-    appendSystemPromptOverride: () => [],
-  });
-  await resourceLoader.reload();
-
-  const { session, modelFallbackMessage } = await createAgentSession({
-    cwd: options.cwd,
-    agentDir: options.configDir,
-    authStorage,
-    modelRegistry,
-    model,
-    thinkingLevel: "medium",
-    tools: READ_ONLY_TOOLS,
-    resourceLoader,
-    sessionManager: SessionManager.create(options.cwd, path.join(options.configDir, "sessions")),
-    settingsManager,
-  });
-
-  const rawHeader = [
-    modelFallbackMessage ? `[pr-review-advisor] ${modelFallbackMessage}` : undefined,
-    `[pr-review-advisor] model=${model.provider}/${model.id}`,
-    `[pr-review-advisor] tools=${READ_ONLY_TOOLS.join(",")}`,
-    "--- ASSISTANT TEXT ---",
-  ].filter((line): line is string => Boolean(line));
-
-  const text = new CappedBuffer(options.maxCaptureBytes);
-  const raw = new CappedBuffer(options.maxCaptureBytes, `${rawHeader.join("\n")}\n`);
-
-  const unsubscribe = session.subscribe((event: AgentSessionEvent) => {
-    if (event.type === "message_update" && event.assistantMessageEvent.type === "text_delta") {
-      text.append(event.assistantMessageEvent.delta);
-      raw.append(event.assistantMessageEvent.delta);
-      return;
-    }
-    if (event.type === "tool_execution_start") {
-      raw.append(`\n[pr-review-advisor] tool_start ${event.toolName}\n`);
-      return;
-    }
-    if (event.type === "tool_execution_end") {
-      raw.append(`[pr-review-advisor] tool_end ${event.toolName} ${event.isError ? "error" : "ok"}\n`);
-      return;
-    }
-    if (event.type === "auto_retry_start") {
-      raw.append(`[pr-review-advisor] retry ${event.attempt}/${event.maxAttempts}: ${event.errorMessage}\n`);
-    }
-  });
-
-  const startedAt = Date.now();
-  const heartbeat = setInterval(() => {
-    const elapsedSeconds = Math.round((Date.now() - startedAt) / 1000);
-    logProgress(`Advisor SDK still running: elapsed=${elapsedSeconds}s timeout=${Math.round(options.timeoutMs / 1000)}s`);
-  }, Math.max(options.heartbeatMs, 1000));
-  heartbeat.unref?.();
-
-  let timeout: NodeJS.Timeout | undefined;
-  const timeoutPromise = new Promise<never>((_resolve, reject) => {
-    timeout = setTimeout(() => {
-      logProgress(`Advisor SDK exceeded timeoutMs=${options.timeoutMs}; aborting session`);
-      void session.abort();
-      reject(new Error(`timed out after ${options.timeoutMs} ms`));
-    }, options.timeoutMs);
-    timeout.unref?.();
-  });
-
-  try {
-    await Promise.race([session.prompt(options.prompt), timeoutPromise]);
-  } finally {
-    unsubscribe();
-    clearInterval(heartbeat);
-    if (timeout) clearTimeout(timeout);
-    try {
-      const exportedPath = await session.exportToHtml(options.htmlExportPath);
-      raw.append(`\n[pr-review-advisor] exported_session_html=${exportedPath}\n`);
-    } catch (error: unknown) {
-      const reason = error instanceof Error ? error.message : String(error);
-      raw.append(`\n[pr-review-advisor] failed_to_export_session_html=${reason}\n`);
-    }
-    session.dispose();
-  }
-
-  const truncationNotes: string[] = [];
-  if (text.droppedBytes > 0) truncationNotes.push(`<assistant text truncated; dropped ${text.droppedBytes} byte(s)>`);
-  if (raw.droppedBytes > 0) truncationNotes.push(`<raw output truncated; dropped ${raw.droppedBytes} byte(s)>`);
-  if (truncationNotes.length > 0) raw.appendFooter(`\n${truncationNotes.join("\n")}\n`);
-
-  return { text: text.toString(), raw: raw.toStringWithTrailingNewline() };
-}
-
-class CappedBuffer {
-  private readonly maxBytes: number;
-  private value: string;
-  public droppedBytes = 0;
-
-  constructor(maxBytes: number, initialValue = "") {
-    this.maxBytes = maxBytes;
-    this.value = initialValue;
-    this.trimToMaxBytes();
-  }
-
-  append(chunk: string): void {
-    this.value += chunk;
-    this.trimToMaxBytes();
-  }
-
-  appendFooter(footer: string): void {
-    const footerBytes = Buffer.byteLength(footer, "utf8");
-    if (footerBytes >= this.maxBytes) {
-      this.value = trimHeadToBytes(footer, this.maxBytes);
-      return;
-    }
-    this.trimToMaxBytes(this.maxBytes - footerBytes);
-    this.value += footer;
-  }
-
-  toString(): string {
-    return this.value;
-  }
-
-  toStringWithTrailingNewline(): string {
-    return this.value.endsWith("\n") ? this.value : `${this.value}\n`;
-  }
-
-  private trimToMaxBytes(maxBytes = this.maxBytes): void {
-    if (Buffer.byteLength(this.value, "utf8") <= maxBytes) return;
-    const trimmed = trimHeadToBytes(this.value, maxBytes);
-    this.droppedBytes += Buffer.byteLength(this.value.slice(0, this.value.length - trimmed.length), "utf8");
-    this.value = trimmed;
-  }
-}
-
-function trimHeadToBytes(value: string, maxBytes: number): string {
-  let removeChars = Math.min(value.length, Math.max(1, Buffer.byteLength(value, "utf8") - maxBytes));
-  while (removeChars < value.length && Buffer.byteLength(value.slice(removeChars), "utf8") > maxBytes) {
-    removeChars += 1;
-  }
-  return value.slice(removeChars);
-}
-
-function parseArgs(argv: string[]): ParsedArgs {
-  const parsed: ParsedArgs = {};
-  for (let i = 0; i < argv.length; i += 1) {
-    const arg = argv[i];
-    if (arg.startsWith("--")) {
-      const key = arg.slice(2).replace(/-([a-z])/g, (_, char: string) => char.toUpperCase());
-      const next = argv[i + 1];
-      if (!next || next.startsWith("--")) {
-        parsed[key] = undefined;
-        continue;
-      }
-      parsed[key] = next;
-      i += 1;
-    }
-  }
-  return parsed;
-}
-
-function readJson<T>(relativeOrAbsolutePath: string): T {
-  return JSON.parse(fs.readFileSync(path.resolve(root, relativeOrAbsolutePath), "utf8")) as T;
-}
-
-function getChangedFiles(base: string, head: string): string[] {
-  const stdout = gitOutput(
-    [
-      ["diff", "--name-only", `${base}...${head}`],
-      ["diff", "--name-only", `${base}..${head}`],
-    ],
-    10 * 1024 * 1024,
-  );
-  if (stdout === undefined) {
-    throw new Error(`failed to diff ${base}..${head}; ensure both refs are fetched`);
-  }
-  return stdout
-    .split(/\r?\n/)
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .sort();
-}
-
-function getDiff(base: string, head: string, maxChars: number): string {
-  const stdout = gitOutput(
-    [
-      ["diff", "--find-renames", "--find-copies", "--unified=80", `${base}...${head}`],
-      ["diff", "--find-renames", "--find-copies", "--unified=80", `${base}..${head}`],
-    ],
-    20 * 1024 * 1024,
-  );
-  return stdout === undefined ? "" : truncate(stdout, maxChars);
-}
-
-function getDiffStat(base: string, head: string): string {
-  return gitOutput(
-    [
-      ["diff", "--stat", `${base}...${head}`],
-      ["diff", "--stat", `${base}..${head}`],
-    ],
-    1024 * 1024,
-  )?.trim() || "<diff stat unavailable>";
-}
-
-function getCommits(base: string, head: string): string[] {
-  return (gitOutput([["log", "--oneline", `${base}..${head}`]], 1024 * 1024) || "")
-    .split(/\r?\n/)
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .slice(0, 50);
-}
-
-function getHeadSha(head: string): string {
-  return execFileSync("git", ["rev-parse", head], { encoding: "utf8" }).trim();
-}
-
-function gitOutput(commands: string[][], maxBuffer: number): string | undefined {
-  for (const command of commands) {
-    try {
-      return execFileSync("git", command, { encoding: "utf8", maxBuffer });
-    } catch {
-      // Try the next form. Some checkouts do not have a merge base locally.
-    }
-  }
-  return undefined;
 }
 
 async function collectDeterministicContext(options: {
@@ -818,42 +515,6 @@ function extractIssueRefs(text: string, prNumber: number): number[] {
   return [...numbers].sort((a, b) => a - b);
 }
 
-async function githubRest<T>(apiPath: string, token: string): Promise<T> {
-  const response = await fetch(`https://api.github.com/${apiPath}`, {
-    headers: {
-      Accept: "application/vnd.github+json",
-      Authorization: `Bearer ${token}`,
-      "X-GitHub-Api-Version": "2022-11-28",
-    },
-  });
-  if (!response.ok) throw new Error(`GitHub REST ${apiPath} failed: ${response.status} ${await response.text()}`);
-  return await response.json() as T;
-}
-
-async function githubRestPaginated<T>(apiPath: string, token: string, limit: number): Promise<T[]> {
-  const results: T[] = [];
-  for (let page = 1; results.length < limit; page += 1) {
-    const separator = apiPath.includes("?") ? "&" : "?";
-    const items = await githubRest<T[]>(`${apiPath}${separator}per_page=${Math.min(100, limit - results.length)}&page=${page}`, token);
-    results.push(...items);
-    if (items.length < 100) break;
-  }
-  return results;
-}
-
-async function githubGraphql(token: string, query: string, variables: Record<string, unknown>): Promise<unknown> {
-  const response = await fetch("https://api.github.com/graphql", {
-    method: "POST",
-    headers: {
-      Accept: "application/vnd.github+json",
-      Authorization: `Bearer ${token}`,
-    },
-    body: JSON.stringify({ query, variables }),
-  });
-  if (!response.ok) throw new Error(`GitHub GraphQL failed: ${response.status} ${await response.text()}`);
-  return await response.json();
-}
-
 function buildPrGraphqlQuery(): string {
   return `
 query($owner: String!, $name: String!, $number: Int!) {
@@ -932,38 +593,6 @@ Git diff, truncated if large:
 ${diff || "<no diff available>"}
 \`\`\`
 `;
-}
-
-function extractJson(text: string, rawPath: string): unknown {
-  const trimmed = text.trim();
-  const candidates = [trimmed, fenced(trimmed), tagged(trimmed, "pr_review_advisor_json"), balancedObject(trimmed)].filter(
-    (candidate): candidate is string => Boolean(candidate),
-  );
-  for (const candidate of candidates) {
-    try {
-      return JSON.parse(candidate);
-    } catch {
-      // Try next candidate.
-    }
-  }
-  throw new Error(`Could not parse JSON from PR review advisor output; see ${rawPath}`);
-}
-
-function fenced(text: string): string | undefined {
-  const match = text.match(/```(?:json)?\s*([\s\S]*?)```/i);
-  return match?.[1]?.trim();
-}
-
-function tagged(text: string, tag: string): string | undefined {
-  const match = text.match(new RegExp(`<${tag}>\\s*([\\s\\S]*?)\\s*</${tag}>`, "i"));
-  return match?.[1]?.trim();
-}
-
-function balancedObject(text: string): string | undefined {
-  const start = text.indexOf("{");
-  const end = text.lastIndexOf("}");
-  if (start === -1 || end === -1 || end <= start) return undefined;
-  return text.slice(start, end + 1);
 }
 
 export function normalizeReviewResult(result: unknown, metadata: ReviewMetadata): ReviewAdvisorResult {
@@ -1159,17 +788,6 @@ export function formatRecommendation(recommendation: SummaryRecommendation): str
   return recommendation.replaceAll("_", " ");
 }
 
-function prepareAdvisorConfig(): { authStorage: AuthStorage; modelRegistry: ModelRegistry } {
-  const authStorage = AuthStorage.inMemory();
-  const modelRegistry = ModelRegistry.inMemory(authStorage);
-  const apiKey = process.env.PR_REVIEW_ADVISOR_API_KEY || process.env.OPENAI_API_KEY;
-  if (apiKey) {
-    authStorage.setRuntimeApiKey(ADVISOR_PROVIDER, apiKey);
-    modelRegistry.registerProvider(ADVISOR_PROVIDER, ADVISOR_PROVIDER_CONFIG);
-  }
-  return { authStorage, modelRegistry };
-}
-
 function unavailableResult(metadata: ReviewMetadata, reason: string, failed: boolean): ReviewAdvisorResult {
   return {
     version: 1,
@@ -1209,47 +827,4 @@ function unavailableResult(metadata: ReviewMetadata, reason: string, failed: boo
       requiresHumanReview: true,
     },
   };
-}
-
-function enumValue<T extends readonly string[]>(value: unknown, allowed: T, fallback: T[number]): T[number] {
-  return typeof value === "string" && allowed.includes(value) ? value : fallback;
-}
-
-function recordItems(value: unknown): Record<string, unknown>[] {
-  return Array.isArray(value) ? value.filter(isRecord) : [];
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
-function stringArray(value: unknown): string[] {
-  return Array.isArray(value) ? value.filter((item): item is string => typeof item === "string" && item.trim().length > 0).map((item) => item.trim()) : [];
-}
-
-function stringOrUndefined(value: unknown): string | undefined {
-  return typeof value === "string" && value.trim() ? value.trim() : undefined;
-}
-
-function stringOrDefault(value: unknown, fallback: string): string {
-  return stringOrUndefined(value) || fallback;
-}
-
-function getPath<T>(value: unknown, pathParts: (string | number)[]): T | undefined {
-  let current: unknown = value;
-  for (const part of pathParts) {
-    if (typeof part === "number") {
-      if (!Array.isArray(current)) return undefined;
-      current = current[part];
-      continue;
-    }
-    if (!isRecord(current)) return undefined;
-    current = current[part];
-  }
-  return current as T | undefined;
-}
-
-function truncate(text: string, maxChars: number): string {
-  if (text.length <= maxChars) return text;
-  return `${text.slice(0, maxChars)}\n\n<diff truncated at ${maxChars} characters>`;
 }
